requirements.txt
appdirs==1.4.3
defusedxml==0.5.0
dj-database-url==0.4.2
Django==1.10.3
django-allauth==0.31.0
django-cors-headers==2.0.2
django-filter==1.0.2
django-rest-auth==0.9.1
django-tenant-schemas==1.7.0
djangorestframework==3.6.2
gunicorn==19.7.1
Markdown==2.6.8
oauthlib==2.0.2
packaging==16.8
psycopg2==2.7.1
pyparsing==2.1.10
python3-openid==3.1.0
requests==2.13.0
requests-oauthlib==0.8.0
six==1.10.0
virtualenv==15.1.0
whitenoise==3.3.0
